requirements.txt
fastapi
uvicorn
sqlmodel
streamlit
pandas
python-multipart
pydantic-settings
google-api-python-client
google-auth
supabase
requests
loguru